Add some `complex_plot` methods
The ModArg method is there for convenience; it sets the unit disk,
zeroing out values which fall outside of it for plotting as having the
forms / functions return zero for q ≥ 1 is technically incorrect and
improper and restricting their domain is impractical.

The tau methods are necessary to resolve dispatch ambiguities since `f`
is generic at ComplexColor. I could just dispatch with an extra ModArg
argument to avoid redundancy, but passing along an extra type when
unnecessary is unintuitive and that also leaves unbounded f(q) methods
available; it's a small price to pay to keep the interface clean and
simple.